refactor `to_casadi` to OOP method
