+++ src/pybamm/__init__.py
@@ -60,7 +60,6 @@ from .expression_tree.operations.evaluate_python import EvaluatorJax
 from .expression_tree.operations.evaluate_python import JaxCooMatrix
 
 from .expression_tree.operations.jacobian import Jacobian
-from .expression_tree.operations.convert_to_casadi import CasadiConverter
 from .expression_tree.operations.unpack_symbols import SymbolUnpacker
 from .expression_tree.operations.serialise import Serialise,ExpressionFunctionParameter
 from .expression_tree.operations.regularise import RegulariseSqrtAndPower
