+++ src/pybamm/expression_tree/binary_operators.py
@@ -182,6 +182,13 @@ class BinaryOperator(pybamm.Symbol):
             f"{self.__class__} does not implement _binary_evaluate."
         )
 
+    def _to_casadi(self, t, y, y_dot, inputs, casadi_symbols):
+        """See :meth:`pybamm.Symbol._to_casadi()`."""
+        converted_left, converted_right = self._children_to_casadi(
+            t, y, y_dot, inputs, casadi_symbols
+        )
+        return self._binary_evaluate(converted_left, converted_right)
+
     def _evaluates_on_edges(self, dimension: str) -> bool:
         """See :meth:`pybamm.Symbol._evaluates_on_edges()`."""
         return self.left.evaluates_on_edges(dimension) or self.right.evaluates_on_edges(
@@ -396,6 +403,13 @@ class KroneckerProduct(BinaryOperator):
         else:
             return np.kron(left, right)
 
+    def _to_casadi(self, t, y, y_dot, inputs, casadi_symbols):
+        """See :meth:`pybamm.Symbol._to_casadi()`."""
+        converted_left, converted_right = self._children_to_casadi(
+            t, y, y_dot, inputs, casadi_symbols
+        )
+        return casadi.kron(converted_left, converted_right)
+
     def _sympy_operator(self, left, right):
         """Override :meth:`pybamm.BinaryOperator._sympy_operator`"""
         raise NotImplementedError
@@ -511,6 +525,20 @@ class MatrixMultiplication(BinaryOperator):
     def _binary_evaluate(self, left, right):
         """See :meth:`pybamm.BinaryOperator._binary_evaluate()`."""
         return left @ right
+
+    def _to_casadi(self, t, y, y_dot, inputs, casadi_symbols):
+        """See :meth:`pybamm.Symbol._to_casadi()`."""
+        from pybamm.expression_tree.operations._casadi_matmul import (
+            try_repeated_row_matmul,
+        )
+
+        converted_left, converted_right = self._children_to_casadi(
+            t, y, y_dot, inputs, casadi_symbols
+        )
+        result = try_repeated_row_matmul(self.children[0], converted_right)
+        if result is not None:
+            return result
+        return self._binary_evaluate(converted_left, converted_right)
 
     def _sympy_operator(self, left, right):
         """Override :meth:`pybamm.BinaryOperator._sympy_operator`"""
@@ -857,6 +885,13 @@ class Modulo(BinaryOperator):
         """See :meth:`pybamm.BinaryOperator._binary_evaluate()`."""
         return left % right
 
+    def _to_casadi(self, t, y, y_dot, inputs, casadi_symbols):
+        """See :meth:`pybamm.Symbol._to_casadi()`."""
+        converted_left, converted_right = self._children_to_casadi(
+            t, y, y_dot, inputs, casadi_symbols
+        )
+        return casadi.fmod(converted_left, converted_right)
+
     def _t_discon(self, expr, y0, inputs, num_events):
         value = expr.evaluate(0, y0, inputs=inputs)
         t_discon = []
@@ -899,6 +934,13 @@ class Minimum(BinaryOperator):
         """See :meth:`pybamm.BinaryOperator._binary_evaluate()`."""
         # don't raise RuntimeWarning for NaNs
         return np.minimum(left, right)
+
+    def _to_casadi(self, t, y, y_dot, inputs, casadi_symbols):
+        """See :meth:`pybamm.Symbol._to_casadi()`."""
+        converted_left, converted_right = self._children_to_casadi(
+            t, y, y_dot, inputs, casadi_symbols
+        )
+        return casadi.fmin(converted_left, converted_right)
 
     def _binary_new_copy(
         self,
@@ -944,6 +986,13 @@ class Maximum(BinaryOperator):
         # don't raise RuntimeWarning for NaNs
         return np.maximum(left, right)
 
+    def _to_casadi(self, t, y, y_dot, inputs, casadi_symbols):
+        """See :meth:`pybamm.Symbol._to_casadi()`."""
+        converted_left, converted_right = self._children_to_casadi(
+            t, y, y_dot, inputs, casadi_symbols
+        )
+        return casadi.fmax(converted_left, converted_right)
+
     def _binary_new_copy(
         self,
         left: ChildSymbol,
@@ -986,6 +1035,13 @@ class Hypot(BinaryOperator):
     def _binary_evaluate(self, left, right):
         """See :meth:`pybamm.BinaryOperator._binary_evaluate()`."""
         return np.hypot(left, right)
+
+    def _to_casadi(self, t, y, y_dot, inputs, casadi_symbols):
+        """See :meth:`pybamm.Symbol._to_casadi()`."""
+        converted_left, converted_right = self._children_to_casadi(
+            t, y, y_dot, inputs, casadi_symbols
+        )
+        return casadi.hypot(converted_left, converted_right)
 
     def _binary_new_copy(
         self,
