+++ src/pybamm/expression_tree/operations/__init__.py
@@ -1,2 +1,2 @@
-__all__ = ['convert_to_casadi', 'evaluate_python', 'jacobian', 'latexify',
+__all__ = ['evaluate_python', 'jacobian', 'latexify',
            'regularise', 'serialise', 'unpack_symbols']
